drf-search/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tqdm==4.67.1
unicode==2.9
Unidecode==1.3.8
uritemplate==4.1.1
